# math/big: improve performance of Binomial